feat: add cjsToEsm solution
